Get PostCSS working in Atlantis
