+++ package.json
@@ -54,6 +54,7 @@
     "@types/lodash": "^4.14.149",
     "@types/react": "^16.9.9",
     "@types/react-dom": "^16.9.2",
+    "autoprefixer": "^9.8.4",
     "css-loader": "^3.4.2",
     "docz": "^2.3.0-alpha.13",
     "docz-core": "^2.3.0-alpha.12",
@@ -71,8 +72,9 @@
     "lodash": "^4.17.15",
     "plop": "^2.5.3",
     "postcss": "^7.0.17",
+    "postcss-import": "^12.0.1",
     "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.6.0",
+    "postcss-preset-env": "^6.7.0",
     "prettier": "^1.18.2",
     "prop-types": "^15.7.2",
     "react": "^16.9.0",
